+++ src/__tests__/integration/event-management.integration.spec.tsx
@@ -10,9 +10,9 @@ import {
   setupMockHandlerCreation,
   setupMockHandlerDeletion,
   setupMockHandlerUpdating,
-} from '../__mocks__/handlersUtils';
-import App from '../App';
-import { server } from '../setupTests';
+} from '../../__mocks__/handlersUtils';
+import App from '../../App';
+import { server } from '../../setupTests';
 import { Event } from '../types';
 
 const theme = createTheme();
