+++ src/__tests__/unit/apis/fetch-holidays.spec.ts
@@ -1,4 +1,4 @@
-import { fetchHolidays } from '../../apis/fetchHolidays';
+import { fetchHolidays } from '../../../apis/fetchHolidays';
 
 describe('fetchHolidays', () => {
   it('주어진 월의 공휴일만 반환한다', () => {
