+++ src/__tests__/unit/utils/date-utils.spec.ts
@@ -1,4 +1,4 @@
-import { Event } from '../../types';
+import { Event } from '../../../types';
 import {
   fillZero,
   formatDate,
@@ -9,7 +9,7 @@ import {
   getWeekDates,
   getWeeksAtMonth,
   isDateInRange,
-} from '../../utils/dateUtils';
+} from '../../../utils/dateUtils';
 
 describe('getDaysInMonth', () => {
   it('1월은 31일 수를 반환한다', () => {
