+++ src/__tests__/unit/utils/event-overlap.spec.ts
@@ -1,10 +1,10 @@
-import { Event } from '../../types';
+import { Event } from '../../../types';
 import {
   convertEventToDateRange,
   findOverlappingEvents,
   isOverlapping,
   parseDateTime,
-} from '../../utils/eventOverlap';
+} from '../../../utils/eventOverlap';
 
 describe('parseDateTime', () => {
   it('2025-07-01 14:30을 정확한 Date 객체로 변환한다', () => {
